worker

+++ mw431552/algorithm.cpp
@@ -13,7 +13,7 @@ std::string another_output_path(const std::string& output_file) {
     }
 }
 
-// void worker(const std::vector<std::vector<float>>& graph, int num_iter, double alpha, double beta, double evaporate, int seed, std::string output_file) {
+// void worker(const std::vector<std::vector<float>>& graph, int num_iter, float alpha, float beta, float evaporate, int seed, std::string output_file) {
 //     std::cout << "Running WORKER algorithm...\n";
 //     std::cout << "Parameters: NUM_ITER=" << num_iter << ", ALPHA=" << alpha 
 //               << ", BETA=" << beta << ", EVAPORATE=" << evaporate << ", SEED=" << seed << std::endl;
@@ -33,7 +33,7 @@ std::string another_output_path(const std::string& output_file) {
 //     ofs.close();
 // }
 
-void queen(const std::vector<std::vector<float>>& graph, int num_iter, double alpha, double beta, double evaporate, int seed, std::string output_file) {
+void queen(const std::vector<std::vector<float>>& graph, int num_iter, float alpha, float beta, float evaporate, int seed, std::string output_file) {
     std::cout << "Running QUEEN algorithm...\n";
     std::cout << "Parameters: NUM_ITER=" << num_iter << ", ALPHA=" << alpha 
               << ", BETA=" << beta << ", EVAPORATE=" << evaporate << ", SEED=" << seed << std::endl;
